Index, Task Form, Task list Views
I changed up the index view. I added in a task form view. The view does not have any back end set up for it yet. That is waiting on Michael to create the models. The Task list view is created with 4 quadrants built into it. It does not have any backend functionality into it yet.